refactor: topics are now fetched from the database instead of JSON file
This commit refactors the storage.js file to retrieve `topics` data from the database rather than from static JSON files. By doing this, we ensure that data stored in localStorage is always up to date with the database. Stored subjects and their nested topics in a single item in localStorage to simplify data management allowing us to filter out a specific subject along with its topics based on its slug.